+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Bug Fixes
 
-- [#431](https://github.com/okta/okta-auth-js/pull/431) Skip non parsable iframe messages for `sdk.fingerprint`
+- [#431](https://github.com/okta/okta-auth-js/pull/431) Skips non parsable iframe messages for `sdk.fingerprint`
 
 ## 3.2.0